matrimony_ platform/matrimony_platform/requirements.txt
fastapi
uvicorn
numpy
sentence-transformers
faiss-cpu
pydantic